+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add: Tags to VPC Flow Logs, VPC IAM Roles, and SSM IAM Roles
 - Add: Variable for container command to execute in ECS
 - Add: Bastion RHEL9 Golang
+- Add: Bastion RHEL9 Docker Compose Plugin
 - Edit: Moving SSM settings out of VPC to its own module to prevent conflicts
 - Edit: Terraform State Bucket Name tag and deletion protection
 - Edit: Changing default Bastion machine type from "t3" to "t3a" for small cost savings